requirements.txt
celery==4.3.0
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.7.1
bcrypt==3.1.7
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
cx-Oracle==7.1.3
cycler==0.10.0
deap==1.3.0
decorator==4.4.0
Django==2.2.2
fabric==2.4.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.21.1
grpcio-reflection==1.21.1
grpcio-tools==1.21.1
h5py==2.9.0
idna==2.8
invoke==1.2.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.0
mock==3.0.5
mysql-connector-python==8.0.16
networkx==2.3
numpy==1.16.4
opt-einsum==3.1.0
pandas==0.24.2
paramiko==2.6.0
prettytable==0.7.2
protobuf==3.9.0
pycparser==2.19
PyNaCl==1.3.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
redis==3.2.1
requests==2.22.0
scipy==1.1.0
six==1.12.0
SQLAlchemy==1.3.8
sqlparse==0.3.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
tornado==6.0.3
urllib3==1.25.3
Werkzeug==0.15.4
wrapt==1.11.2
django-cors-headers==3.1.1
djangorestframework==3.10.3
gunicorn==20.0.2
gevent==1.4.0
supervisor==4.1.0
whitenoise==4.1.4
django-redis==4.8.0
coreapi==2.3.3
drf-extensions==0.5.0
whitenoise==4.1.4
mysqlclient==1.3.13